[#50] 잘못된 패키지명 수정
